+++ src/lib.rs
@@ -86,6 +86,7 @@ macro_rules! impl_inherent_methods {
 impl_inherent_methods!(Adx, AdxConfig, AdxConfigBuilder);
 impl_inherent_methods!(Atr, AtrConfig, AtrConfigBuilder);
 impl_inherent_methods!(Cci, CciConfig, CciConfigBuilder);
+impl_inherent_methods!(Chop, ChopConfig, ChopConfigBuilder);
 impl_inherent_methods!(Dc, DcConfig, DcConfigBuilder);
 impl_inherent_methods!(Sma, SmaConfig, SmaConfigBuilder);
 impl_inherent_methods!(Rsi, RsiConfig, RsiConfigBuilder);